+++ test/support/image_pipe/plug_test/consume_large_source_image.ex
@@ -1,0 +1,9 @@
+defmodule ImagePipe.PlugTest.ConsumeLargeSourceImage do
+  def open(stream, decode_options) do
+    _chunks = Enum.to_list(stream)
+
+    "priv/static/images/beach.jpg"
+    |> File.read!()
+    |> Image.open(decode_options)
+  end
+end
