+++ test/support/image_pipe/plug_test/consume_source_then_decode_error_image.ex
@@ -1,0 +1,6 @@
+defmodule ImagePipe.PlugTest.ConsumeSourceThenDecodeErrorImage do
+  def open(stream, _decode_options) do
+    _chunks = Enum.to_list(stream)
+    {:error, :forced_decode_error}
+  end
+end
